This addresses the performance regression of `mapslices` introduced by #27417 and noticed in  #27828. I think replacing `isconcretetype(typea) || isType(typea)` with `unioncomplexity(typea) == 0` makes a lot of sense and is in the same spirit as the original PR.

However, I had to introduce another condition due to this quirk (which happens both before #27417 and with this PR):
```julia
julia> foo(a,b,c,d) = SubArray(a,b,c,d)
foo (generic function with 1 method)

julia> atypes = Tuple{Union{IndexCartesian, IndexLinear}, Any, Union{Tuple{}, Tuple{Any,Vararg{Any,N} where N}}, Tuple}
Tuple{Union{IndexCartesian, IndexLinear},Any,Union{Tuple{}, Tuple{Any,Vararg{Any,N} where N}},Tuple}

julia> atypes2 = Tuple{Union{IndexCartesian, IndexLinear}, Any, Tuple, Tuple}
Tuple{Union{IndexCartesian, IndexLinear},Any,Tuple,Tuple}

julia> atypes <: atypes2
true

julia> code_typed(foo, atypes)
1-element Array{Any,1}:
 CodeInfo(
1 1 ─ %1 = Main.SubArray(%%a, %%b, %%c, %%d)::SubArray                      │
  └──      return %1                                                        │
) => SubArray

julia> code_typed(foo, atypes2)
1-element Array{Any,1}:
 CodeInfo(
1 1 ─ %1 = Main.SubArray(%%a, %%b, %%c, %%d)::Union{SubArray{_1,_2,_3,_4,false} where _4 where _3 where _2 where _1, SubArray{_1,_2,_3,_4,true} where _4 where _3 where _2 where _1}
  └──      return %1                                                        │
) => Union{SubArray{_1,_2,_3,_4,false} where _4 where _3 where _2 where _1, SubArray{_1,_2,_3,_4,true} where _4 where _3 where _2 where _1}
```
Note that the wider `atypes2` gives the narrower inference result. Hence the additional `Tuple` condition to turn `Union{Tuple{}, Tuple{Any,Vararg{Any,N} where N}}` into `Tuple`. (They should probably be type-equal anyway, but that's a different issue.) This leaves me a bit unsatisfied, but seems to be necessary to fix the regression.