@JeffBezanson 2018-06-28
A non-monotonicity is pretty bad; I think it would be better to fix that and address the regression by improving the inlining logic (which I'm working on now).