@martinholters 2018-06-29
I've looked a little bit into the non-montonicity. Union-splitting the `atypes` case gives these four combinations:

argtypes | return type
--- | ---
`Tuple{IndexCartesian,Any,Tuple{},Tuple}` | `SubArray{_1,_2,_3,Tuple{},false} where _3 where _2 where _1`
`Tuple{IndexLinear,Any,Tuple{},Tuple}` | `SubArray{_1,_2,_3,Tuple{},true} where _3 where _2 where _1`
`Tuple{IndexCartesian,Any,Tuple{Any,Vararg{Any,N} where N},Tuple}` | `SubArray{_1,_2,_3,_4,false} where _4 where _3 where _2 where _1`
`Tuple{IndexLinear,Any,Tuple{Any,Vararg{Any,N} where N},Tuple}` | `SubArray{_1,_2,_3,_4,true} where _4 where _3 where _2 where _1`

Then `tmerge`'ing the first two gives a `Union` and then `tmerge`'ing the third widens to `SubArray`. On the other hand, with `atypes2`, union-splitting only gives two combinations:

argtypes | return type
--- | ---
`Tuple{IndexCartesian,Any,Tuple,Tuple}` | `SubArray{_1,_2,_3,_4,false} where _4 where _3 where _2 where _1`
`Tuple{IndexLinear,Any,Tuple,Tuple}` | `SubArray{_1,_2,_3,_4,true} where _4 where _3 where _2 where _1`

These are `tmerge`d into the `Union`, done.

Now interestingly, if the order of the `tmerge` operations in the first case was changed, one could obtain the same results, as the first return type is a subtype of the third and the second one is a subtype of the fourth. I have an idea how to make `tmerge` also give a `Union` in this case. I'll give it a try. Meanwhile, #27857 might be the better fix to address the issue that brought me here.