@martinholters 2018-06-29
Ok, this is now much more elegant. However, as the original issue is better tackled with #27857 (addressing the underlying problem instead of working around it), the need for this is much lower. It may still give slightly better inference results, but whether they give tangible benefits is another question. It does fix above-mentioned quirk, though, as now
```julia
julia> code_typed(foo, atypes)
1-element Array{Any,1}:
 CodeInfo(
1 1 ─ %1 = Main.SubArray(%%a, %%b, %%c, %%d)::Union{SubArray{_1,_2,_3,_4,false} where _4 where _3 where _2 where _1, SubArray{_1,_2,_3,_4,true} where _4 where _3 where _2 where _1}
  └──      return %1                                                        │
) => Union{SubArray{_1,_2,_3,_4,false} where _4 where _3 where _2 where _1, SubArray{_1,_2,_3,_4,true} where _4 where _3 where _2 where _1}

julia> code_typed(foo, atypes2)
1-element Array{Any,1}:
 CodeInfo(
1 1 ─ %1 = Main.SubArray(%%a, %%b, %%c, %%d)::Union{SubArray{_1,_2,_3,_4,false} where _4 where _3 where _2 where _1, SubArray{_1,_2,_3,_4,true} where _4 where _3 where _2 where _1}
  └──      return %1                                                        │
) => Union{SubArray{_1,_2,_3,_4,false} where _4 where _3 where _2 where _1, SubArray{_1,_2,_3,_4,true} where _4 where _3 where _2 where _1}
```