@martinholters 2018-07-02
@nanosoldier `runbenchmarks(ALL, vs=":master")`