@nanosoldier 2018-07-02
[Your benchmark job](https://github.com/JuliaLang/julia/pull/27843#issuecomment-401688753) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/3cf14bbe766eb97dcd39ab1e33a3359c3a8c5aac/b079a9b_vs_b7dc60f/report.md). cc @ararslan