@martinholters 2018-07-11
@nanosoldier `runbenchmarks(ALL, vs=":master")`