@martinholters 2018-07-11
@ararslan is the soldier down? I wanted to check how the recent changes, in particular wrt inlining, interact with this PR.