@nanosoldier 2018-07-11
[Your benchmark job](https://github.com/JuliaLang/julia/pull/27843#issuecomment-404267313) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/f7fa39e86600f65b8ea2e0c950c720c409a4682b/12618d2_vs_3dad306/report.md). cc @ararslan