@StefanKarpinski 2018-07-12
It does seem nicer to me as well in my very mildly informed opinion. @JeffBezanson?