@JeffBezanson 2018-07-12
I would worry about compiler performance here; using more Union types will cause us to infer more signatures.