@JeffBezanson 2018-07-12
If we're going to merge this it would be good to add some sort of test related to https://github.com/JuliaLang/julia/pull/27843#issuecomment-401295448.