@martinholters 2018-07-13
Not sure how good a benchmark of compiler performance building the sysmimg is, but:
master:
```
Total ─────── 140.776988 seconds 
Base: ───────  32.209240 seconds 22.8796%
Stdlibs: ────  92.753201 seconds 65.8866%
Precompile: ─  15.812288 seconds 11.2322%
```
this PR:
```
Total ─────── 147.342998 seconds 
Base: ───────  34.227599 seconds 23.2299%
Stdlibs: ────  97.176708 seconds 65.9527%
Precompile: ─  15.936632 seconds 10.816%
```
So, yes, might cost us a few percent. 

Test to the same effect as OP, but not relying on the specifics of the `SubArray` constructor added.