@vtjnash 2020-12-28
They both give `SubArray` now, but apparently I said this is good, if you care to rebase