@martinholters 2021-01-12
Added a test to address https://github.com/JuliaLang/julia/pull/27843#issuecomment-404676639. CI had been all green, now there is a fail in SHA on tester_linux64 which I don't understand but assume unrelated. If someone can confirm it's unrelated to this PR, I'll go ahead and merge unless anyone objects.