@vtjnash 2021-01-12
It is an AMD cpu bug